+++ cont_curent.cpp
@@ -7,7 +7,7 @@
 #include "exceptii.h"
 
 
-contcurent::contcurent(int suma_, const std::string &moneda_, const std::string &iban_,Titular* titular_, int c,float abonament_telefon_, float utilitati_) : Cont( suma_,  moneda_,  iban_, titular_, c) ,abonament_telefon(abonament_telefon_), utilitati(utilitati_){
+contcurent::contcurent(float suma_, const std::string &moneda_, const std::string &iban_,Titular* titular_, int c,float abonament_telefon_, float utilitati_) : Cont( suma_,  moneda_,  iban_, titular_, c) ,abonament_telefon(abonament_telefon_), utilitati(utilitati_){
 
     if(suma_< suma_min) throw(eroare_suma_cont{"Suma insuficienta\n"});
     else
